+++ server/src/main/java/com/example/hackathon/account/domain/AccountCreationCommand.java
@@ -1,4 +1,4 @@
-package com.example.hackathon.account.interfaces;
+package com.example.hackathon.account.domain;
 
 public class AccountCreationCommand {
     private String accountTypeUniqueNo;
