+++ server/src/main/java/com/example/hackathon/account/domain/SHAccountStorePort.java
@@ -1,7 +1,6 @@
 package com.example.hackathon.account.domain;
 
 import com.example.hackathon.account.domain.shResponse.SHAccountCreationREC;
-import com.example.hackathon.account.interfaces.AccountCreationCommand;
 
 public interface SHAccountStorePort {
     SHAccountCreationREC createAccountSummary(AccountCreationCommand command);
